+++ cmd/routes.go
@@ -6,6 +6,7 @@ package cmd
 
 import (
 	"github.com/zhuharev/hyip/models"
+	"github.com/zhuharev/hyip/web/context"
 	"github.com/zhuharev/hyip/web/routers/auth"
 	"github.com/zhuharev/hyip/web/routers/dash"
 	"github.com/zhuharev/hyip/web/routers/pages"
@@ -23,23 +24,29 @@ func registreRoutes(m *macaron.Macaron) {
 	m.Get("/about/:slug", pages.Show)
 	m.Post("/about/:slug", pages.Update)
 	m.Any("/about/:slug/edit", pages.Edit)
-	m.Post("/dash/settings/advcash", dash.MakeSaveField("Advcash"))
-	//m.Get("/", pages.Index)
+
+	m.Get("/", context.Toggle(&context.ToggleOptions{SignOutRequired: true}), pages.Index)
 	m.Post("/tools/calc", tools.Calc)
-	m.Get("/dash", dash.Index)
-	m.Get("/dash/partners", dash.Partners)
-	m.Get("/dash/contracts", dash.Contracts)
-	m.Get("/dash/transactions", dash.Transactions)
-	m.Get("/dash/settings", dash.Settings)
-	m.Post("/dash/settings", dash.UpdateSettings)
-	m.Post("/dash/settings/plans/create", plans.Create)
+
+	m.Group("/dash", func() {
+		m.Get("/", dash.Index)
+		m.Get("/partners", dash.Partners)
+		m.Get("/contracts", dash.Contracts)
+		m.Get("/transactions", dash.Transactions)
+		m.Get("/settings", dash.Settings)
+		m.Post("/settings", dash.UpdateSettings)
+		m.Post("/settings/plans/create", plans.Create)
+		m.Post("/settings/advcash", dash.MakeSaveField("Advcash"))
+	}, context.Toggle(&context.ToggleOptions{SignInRequired: true}))
+
 	m.Get("/support", support.Index)
 	m.Post("/support/tickets/create", support.CreateTicket)
 	m.Get("/support/tickets/:ticketID", support.Ticket)
 	m.Post("/support/messages/send", support.SendMessage)
-	m.Any("/login", auth.Login)
-	m.Any("/reg", auth.Reg)
-	m.Get("/logout", auth.Logout)
+	m.Get("/support/admin", support.Admin)
+	m.Any("/login", context.Toggle(&context.ToggleOptions{SignOutRequired: true}), auth.Login)
+	m.Any("/reg", context.Toggle(&context.ToggleOptions{SignOutRequired: true}), auth.Reg)
+	m.Get("/logout", context.Toggle(&context.ToggleOptions{SignInRequired: true}), auth.Logout)
 
 	m.Post("/external/webhooks/qiwi", binding.Bind(models.QiwiTxn{}), webhooks.Qiwi)
 }