+++ web/routers/support/admin.go
@@ -2,16 +2,20 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-package pages
+package support
 
 import (
 	"github.com/zhuharev/hyip/models"
 	"github.com/zhuharev/hyip/web/context"
-	macaron "gopkg.in/macaron.v1"
 )
 
-// Index is pages.index controller
-func Index(ctx *context.Context) {
-	_ = models.User{}
-	ctx.HTML(200, "pages/index", nil, macaron.HTMLOptions{Layout: ""})
+// Admin is support.createticket controller
+func Admin(ctx *context.Context) {
+
+	tickets, err := models.Tickets.AdminList()
+	if ctx.HasError(err) {
+		return
+	}
+	ctx.Data["tickets"] = tickets
+	ctx.HTML(200, "support/admin")
 }