+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.0] 2017-11-30
+## Added
+- Support: tickets and admin.
+- Redirect from login and reg pages when already signed in. Redirect from dash,
+ if signed out
+
 ## [0.0.2] 2017-11-29
 ### Added
 - fixer.io for currency conversation